Generalise time coord fix
