+++ esmvalcore/cmor/_fixes/cordex/ichec_ec_earth/knmi_racmo22e.py
@@ -1,4 +1,4 @@
 """Fixes for rcm KNMI-RACMO22E driven by ICHEC-EC-EARTH."""
-from ..ncc_noresm1_m.gerics_remo2015 import Pr as BaseFix
+from ..cordex_fixes import TimeLongName as BaseFix
 
 Pr = BaseFix
