+++ esmvalcore/cmor/_fixes/cordex/ichec_ec_earth/smhi_rca4.py
@@ -1,5 +1,5 @@
 """Fixes for rcm SMHI-RCA4 driven by ICHEC-EC-EARTH."""
-from ..ncc_noresm1_m.gerics_remo2015 import Pr as BaseFix
+from ..cordex_fixes import TimeLongName as BaseFix
 
 Pr = BaseFix
