+++ esmvalcore/cmor/_fixes/cordex/miroc_miroc5/gerics_remo2015.py
@@ -1,4 +1,4 @@
 """Fixes for rcm GERICS-REMO2015 driven by MIROC-MIROC5."""
-from ..ncc_noresm1_m.gerics_remo2015 import Pr as BaseFix
+from ..cordex_fixes import TimeLongName as BaseFix
 
 Pr = BaseFix
