+++ esmvalcore/cmor/_fixes/cordex/mpi_m_mpi_esm_lr/ictp_regcm4_6.py
@@ -1,5 +1,5 @@
 """Fixes for rcm ICTP-RegCM4-6 driven by MPI-M-MPI-ESM-LR."""
-from ..ncc_noresm1_m.gerics_remo2015 import Pr as BaseFix
+from ..cordex_fixes import TimeLongName as BaseFix
 
 Pr = BaseFix
