+++ esmvalcore/cmor/_fixes/cordex/ncc_noresm1_m/knmi_racmo22e.py
@@ -1,4 +1,4 @@
 """Fixes for rcm KNMI-RACMO22E driven by NCC-NorESM1-M."""
-from .gerics_remo2015 import Pr as BaseFix
+from ..cordex_fixes import TimeLongName as BaseFix
 
 Pr = BaseFix
