+++ esmvalcore/cmor/_fixes/cordex/ncc_noresm1_m/smhi_rca4.py
@@ -1,5 +1,5 @@
 """Fixes for rcm SMHI-RCA4 driven by NCC-NorESM1-M."""
-from .gerics_remo2015 import Pr as BaseFix
+from ..cordex_fixes import TimeLongName as BaseFix
 
 Pr = BaseFix
